chore: move global configuration files to separate folder and update imports accordingly
