+++ src/commands/anniversaire/[sub-commands]/set.cmd.ts
@@ -1,7 +1,7 @@
 import type { CommandExecute } from "$core/utils/handler/command";
 import { DayJS } from "$core/utils/day-js";
 import { commands } from "$core/configs/message/command";
-import { global } from "$core/configs/global.config";
+import { global } from "$core/configs/global";
 import { getAge } from "$core/utils/function";
 import { simpleEmbed } from "$core/utils/embed";
 import { msgParams } from "$core/utils/message";
