+++ src/commands/emoji/emoji.cmd.ts
@@ -4,7 +4,7 @@ import { simpleEmbed } from "$core/utils/embed";
 import { msgParams } from "$core/utils/message";
 import { commands } from "$core/configs/message/command";
 import { AttachmentBuilder, BaseGuildTextChannel } from "discord.js";
-import { proposals } from "$core/configs";
+import { proposals } from "$core/configs/global";
 import { logger } from "$core/utils/logger";
 import { getGuildTypeById, guilds } from "$core/configs/guild";
 import { userWithId } from "$core/utils/user";
