+++ src/commands/role/role.util.ts
@@ -1,5 +1,5 @@
 import { getRolesByCategory } from "$core/api/func/main-role";
-import { interactionId } from "$core/configs";
+import { interactionId } from "$core/configs/global";
 import type { Guild, GuildMember, SelectMenuComponentOptionData } from "discord.js";
 import { ActionRowBuilder, StringSelectMenuBuilder } from "discord.js";
