+++ src/commands/stats/stats.util.ts
@@ -1,9 +1,9 @@
 import type { GraphType } from "./stats.type";
-import { global } from "$core/configs";
-import { DayJS } from "$core/utils/day-js";
-import { dateFormat } from "$core/utils/function";
 import type { GetServerActivityHistoryQuery } from "$core/utils/request/graphql";
 import type { ChartConfiguration } from "chart.js";
+import { global } from "$core/configs/global";
+import { DayJS } from "$core/utils/day-js";
+import { dateFormat } from "$core/utils/function";
 
 export const generateChartConfig = (
   type: GraphType,
