+++ src/configs/global/global.type.ts
