+++ src/configs/global/index.ts
