+++ src/utils/embed.ts
@@ -1,5 +1,5 @@
 import { EmbedBuilder } from "discord.js";
-import { global } from "$core/configs";
+import { global } from "$core/configs/global";
 
 type EmbedType = "normal" | "error";
